pyproject.toml
[tool.poetry]
name = "chipbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
discord = "^1.0.1"
replit = "^2.0.0"
PyNaCl = "^1.4.0"
mutagen = "^1.45.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
